upd2

+++ Lab1C/lab2.cpp
@@ -21,7 +21,7 @@ void task(Stack* stack)
             s += c;
         }
     }
-    for (int i = 0; i <= s.length(); i++)
+    for (size_t i = 0; i <= s.length(); i++)
     {
         if (s[i] == '*')
         {
